+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- This adapter now supports Harlequin's `--read-only` option and declares `IMPLEMENTS_READ_ONLY`. Read-only connections are enforced by the server, using `default_transaction_read_only` ([#58](https://github.com/tconbeer/harlequin-postgres/issues/58)).
+
 ## [1.3.1] - 2026-04-19
 
 - Fixes a bug causing user schemas starting with "pg" to be filtered out of the catalog. (#50)